Move Component style to separate files.
