+++ src/components/Carousel/Carousel.style.js
@@ -1,0 +1,12 @@
+import { makeStyles } from "@material-ui/core";
+
+export default makeStyles(theme => ({
+  carousel: {
+    position: "relative",
+    width: "100%",
+    height: "400px",
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center"
+  }
+}));
