+++ src/components/Carousel/CarouselArrow/CarouselArrow.js
@@ -1,25 +1,11 @@
 import React from "react";
-import { makeStyles, IconButton } from "@material-ui/core";
+import { IconButton } from "@material-ui/core";
 import {
   ArrowLeft as ArrowLeftIcon,
   ArrowRight as ArrowRightIcon
 } from "@material-ui/icons";
 
-const useStyles = makeStyles(theme => ({
-  slideArrow: {
-    color: theme.palette.primary.main,
-    backgroundColor: theme.palette.primary.contrastText,
-    cursor: "pointer",
-    position: "absolute",
-    top: "50%"
-  },
-  left: {
-    left: theme.spacing(1)
-  },
-  right: {
-    right: theme.spacing(1)
-  }
-}));
+import useStyles from "./CarouselArrow.style";
 
 export function CarouselArrow({ direction, clickFunction }) {
   const classes = useStyles();
